llm-embedding/requirements.txt
faiss-cpu
sentence-transformers
torch
hf_xet
huggingface_hub[hf_xet]
chromadb

